Moved RangeMiddleware to cobspec
